Makes JacksonInnerValues package private inner classes of ValueImpls
