+++ wdtk-datamodel/src/main/java/org/wikidata/wdtk/datamodel/implementation/GlobeCoordinatesValueImpl.java
@@ -20,10 +20,10 @@ package org.wikidata.wdtk.datamodel.implementation;
  * #L%
  */
 
+import org.apache.commons.lang3.Validate;
 import org.wikidata.wdtk.datamodel.helpers.Equality;
 import org.wikidata.wdtk.datamodel.helpers.Hash;
 import org.wikidata.wdtk.datamodel.helpers.ToString;
-import org.wikidata.wdtk.datamodel.implementation.json.JacksonInnerGlobeCoordinates;
 import org.wikidata.wdtk.datamodel.interfaces.GlobeCoordinatesValue;
 import org.wikidata.wdtk.datamodel.interfaces.ValueVisitor;
 
@@ -31,7 +31,6 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.databind.JsonDeserializer.None;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 
 /**
@@ -43,7 +42,7 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
  *
  */
 @JsonIgnoreProperties(ignoreUnknown = true)
-@JsonDeserialize(using = None.class)
+@JsonDeserialize()
 public class GlobeCoordinatesValueImpl extends ValueImpl implements
 		GlobeCoordinatesValue {
 
@@ -65,7 +64,7 @@ public class GlobeCoordinatesValueImpl extends ValueImpl implements
 	 * @param globe
 	 *            IRI specifying the celestial objects of the coordinates
 	 */
-	public GlobeCoordinatesValueImpl(double latitude, double longitude,
+	GlobeCoordinatesValueImpl(double latitude, double longitude,
 			double precision, String globe) {
 		super(JSON_VALUE_TYPE_GLOBE_COORDINATES);
 		this.value = new JacksonInnerGlobeCoordinates(latitude, longitude,
@@ -77,7 +76,7 @@ public class GlobeCoordinatesValueImpl extends ValueImpl implements
 	 * Constructor for deserialization from JSON via Jackson.
 	 */
 	@JsonCreator
-	protected GlobeCoordinatesValueImpl(
+	GlobeCoordinatesValueImpl(
 			@JsonProperty("value") JacksonInnerGlobeCoordinates innerCoordinates) {
 		super(JSON_VALUE_TYPE_GLOBE_COORDINATES);
 		this.value = innerCoordinates;
@@ -135,5 +134,92 @@ public class GlobeCoordinatesValueImpl extends ValueImpl implements
 	@Override
 	public String toString() {
 		return ToString.toString(this);
+	}
+
+	/**
+	 * Helper object that represents the JSON object structure of the value.
+	 */
+	@JsonIgnoreProperties(ignoreUnknown = true)
+	static class JacksonInnerGlobeCoordinates {
+
+		private final double latitude;
+		private final double longitude;
+		private final double precision;
+		private final String globe;
+
+		/**
+		 * Constructor. Creates an oject that can be populated during JSON
+		 * deserialization. Should only be used by Jackson for this very purpose.
+		 */
+		@JsonCreator
+		JacksonInnerGlobeCoordinates(
+				@JsonProperty("latitude") double latitude,
+				@JsonProperty("longitude") double longitude,
+				@JsonProperty("precision") double precision,
+				@JsonProperty("globe") String globe)  {
+			Validate.notNull(globe, "globe IRI must not be null");
+			if ((latitude > 90 * PREC_DEGREE)
+					|| (latitude < -90 * PREC_DEGREE)) {
+				throw new IllegalArgumentException(
+						"Latitude must be between 90 degrees and -90 degrees.");
+			}
+			if ((longitude > 360 * PREC_DEGREE)
+					|| (longitude < -360 * PREC_DEGREE)) {
+				throw new IllegalArgumentException(
+						"Longitude must be between -360 degrees and +360 degrees.");
+			}
+			this.latitude = latitude;
+			this.longitude = longitude;
+			if (precision <= 0.0) {
+				// We just do this silently because it is so common in the data.
+				// Precision "0" does not make sense for a physical quantity.
+				// Automatic precision does not make sense for floating point
+				// values. "0" also is commonly produced from "null" in JSON.
+				this.precision = PREC_ARCSECOND;
+			} else {
+				this.precision = precision;
+			}
+			this.globe = globe;
+		}
+
+		/**
+		 * Returns the latitude.
+		 *
+		 * @see GlobeCoordinatesValue#getLatitude()
+		 * @return latitude
+		 */
+		public double getLatitude() {
+			return this.latitude;
+		}
+
+		/**
+		 * Returns the longitude.
+		 *
+		 * @see GlobeCoordinatesValue#getLongitude()
+		 * @return longitude
+		 */
+		public double getLongitude() {
+			return this.longitude;
+		}
+
+		/**
+		 * Returns the precision.
+		 *
+		 * @see GlobeCoordinatesValue#getPrecision()
+		 * @return precision
+		 */
+		public double getPrecision() {
+			return this.precision;
+		}
+
+		/**
+		 * Returns the globe.
+		 *
+		 * @see GlobeCoordinatesValue#getGlobe()
+		 * @return globe
+		 */
+		public String getGlobe() {
+			return this.globe;
+		}
 	}
 }
