+++ wdtk-datamodel/src/main/java/org/wikidata/wdtk/datamodel/implementation/ItemIdValueImpl.java
@@ -3,7 +3,6 @@ package org.wikidata.wdtk.datamodel.implementation;
 import org.wikidata.wdtk.datamodel.helpers.Equality;
 import org.wikidata.wdtk.datamodel.helpers.Hash;
 import org.wikidata.wdtk.datamodel.helpers.ToString;
-import org.wikidata.wdtk.datamodel.implementation.json.JacksonInnerEntityId;
 import org.wikidata.wdtk.datamodel.interfaces.EntityIdValue;
 import org.wikidata.wdtk.datamodel.interfaces.ItemIdValue;
 import org.wikidata.wdtk.datamodel.interfaces.ValueVisitor;
@@ -13,7 +12,6 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.databind.JsonDeserializer.None;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 
 /*
@@ -44,7 +42,7 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
  *
  */
 @JsonIgnoreProperties(ignoreUnknown = true)
-@JsonDeserialize(using = None.class)
+@JsonDeserialize()
 public class ItemIdValueImpl extends EntityIdValueImpl implements
 		ItemIdValue {
 	
@@ -60,7 +58,7 @@ public class ItemIdValueImpl extends EntityIdValueImpl implements
 			String id,
 			String siteIri) {
 		super(id, siteIri);
-		checkEntityIdType();
+		assertHasJsonEntityType(JSON_ENTITY_TYPE_ITEM);
 	}
 	/**
 	 * Constructor used for deserialization with Jackson.
@@ -71,22 +69,11 @@ public class ItemIdValueImpl extends EntityIdValueImpl implements
 	 *     the siteIRI that this value refers to.
 	 */
 	@JsonCreator
-	protected ItemIdValueImpl(
+	ItemIdValueImpl(
 			@JsonProperty("value") JacksonInnerEntityId value,
 			@JacksonInject("siteIri") String siteIri) {
 		super(value, siteIri);
-		checkEntityIdType();
-	}
-	
-	/**
-	 * Checks that the entity id is of the right type.
-	 */
-	private void checkEntityIdType() {
-		if (!JacksonInnerEntityId.JSON_ENTITY_TYPE_ITEM.equals(value
-				.getJsonEntityType())) {
-			throw new RuntimeException("Unexpected inner value type: "
-					+ value.getJsonEntityType());
-		}
+		assertHasJsonEntityType(JSON_ENTITY_TYPE_ITEM);
 	}
 
 	@JsonIgnore
